backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
python-dotenv
supabase
pydantic
PyMySQL
python-jose
